Move props modules out of pages/ directory.
