+++ src/core/js-syntax.js
