+++ src/core/neon-syntax.js
