+++ src/pages/index.js
@@ -5,8 +5,8 @@ import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import useBaseUrl from "@docusaurus/useBaseUrl";
 import SyntaxHighlighter from "react-syntax-highlighter";
 import { Button, ButtonGroup, Row, Col, Container } from "reactstrap";
-import jsSyntax from "./js-syntax";
-import neonSyntax from "./neon-syntax";
+import jsSyntax from "../core/js-syntax";
+import neonSyntax from "../core/neon-syntax";
 import styles from "./styles.module.css";
 import Asciinema from "./asciinema";
 import "../css/custom.css";
